code/api/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.5
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
